refactor(example/anomaly-monitor): モジュールを分割し監視システムを整理
- Main.hsの肥大化した実装をAnomaly, Banner, Config, Cpu, Env, Exception, Loop各モジュールへ分割
- 設定・環境・例外・バナー表示・CPU統計・異常判定・監視ループを個別に整理
- himari.cabalに新モジュールを追加しビルド対象を更新
- 設定表示やエラー処理も各責務ごとに整理し可読性・保守性を向上
