+++ example/anomaly-monitor/Exception.hs
@@ -1,0 +1,11 @@
+module Exception
+  ( AnomalyMonitorException (..)
+  ) where
+
+import Himari
+
+newtype AnomalyMonitorException
+  = ConfigDecodeException Text
+  deriving (Eq, Generic, Show)
+
+instance Exception AnomalyMonitorException
